Animation and refactoring
I had to think about it a bit to add animation to the game (and there are bugs) but now you can run around and run into walls. There were also collision bugs that centered around Rect() being int based instead of float based. I added a custom RectF() class to take care of that. 

So far so good.