Lots of additions
Had some time on my flight back to Baltimore to work on a lot of loose ends. Fixed up collision deteciton some...although there is more work to do there. Also got the random dungeon generation back to where it was on the Uzebox version. Also added collectibles, so now you can pick up the animated treasure chest. Speaking of which, fixed a subtle bug in the animation system that reared it's head once more that 1 Entity in the game was animated.

A big set of changes, but there is a long way to go.