Font!
Did the manual labor of generating the font textures from my tileset, and implemented a routine to allow for text to be displayed in the game. Also updated the DecorateMapTIles() routine to catch end cap pieces.

Also started implementing game states...but I'm sure it'll change a lot in the future.