Fleshed out TextSystem
Added some more functionality to the TextSystem. Now, HasText components can be multi-line, as well as multi-paged. Not sure this is the final implementation of that feature, but we'll see. Still more todo, as usual.

Also, fixed the Z-ordering of things drawn to the screen. Now, an enum in Globals.cs controls what order things are drawn in, and is much easier to maintain over the old arbitrary system.