Changed up Input system
Realizing that my quicky InputSystem implementation was very inflexible, I made an architectural change that will make things much easier going forward. I'm sure I'm still going to have to make more changes in the same area going forward, but for now, the game plays the same as it did before.

Oh...and I added the beginnings of a sword. Nothing to see yet, but it's a start.