Lots of structural changes
So, I had run into a dead end in trying to add more features due to the way the ECS system had been integrated with EfD2. I have played around a bit, and added in the beginnings of an event system to better facilitate compartmentalizing individual systems. I had put in some workarounds before that were preventing me from adding new features, so hopefully this is the start of big changes to come. Some of the stuff I'm checking in might go away, and the game is partially broken right now, but I feel better about this.