Lots, and lots of changes
I added in a TimeSystem to properly handle any Entity that has an Ephemeral component. I redid the way Events are handled by factoring in the GameState, and GameSystem. So now, when passing from the Entrance to Exit of a room, the handling of this is generalized in a way to be useful in the future.

It's a lot of changes, but most of these feel like the right choice going forward.

I was in the middle of adding an Inventory to the player, with the ability to add Collectibles to  it, but ran out of time for now. Will continue on at this point next time.