You can get gold!
I added a Collectible System to allow the player (or others) to pick up items and add them to their inventory. There was a nasty bug in there that was still causing collisions to occur after an item collection, but its fixed now.

Also...added a HUD! You can now see how much gold you've picked up. I'm not entirely happy with how all the Systems are laid out and interplay, but for the moment, things are progressing well. I should be able to add features a little more quickly now...at least hopefully.

I also fixed a positioning bug in the Text System, and generalized it a bit more as well.