Added Health display
I quickly added the display of the Player's health. This isn't the final implementation...and will indeed fail if more than 1 Entity has a health component...but I feel good about adding features.

I also encountered a bug where, when the player hits the exit and is moved to the entrance, they will get stuck because their CollidingEntities is never cleared. It still thinks the player is colliding. Will look into that later.