Refactors...
So, it's been difficult to see exactly how having the player attack is going to fit into this hodge-podge architecture I've been building, so I refactored a bit of code to help angle me in the right direction. In my mind, the Player's current state should be driven off of the ActorStateType, and not be coupled to the AnimationType. The AnimationType will be triggered by the ActorStateType, and thus play correctly when needed.

I'll just use the input to change the ActorStateType to Attacking, and then spawn the weapon from the first equipped weapon slot.  Hrm...that sounds complicated too...