Major refactor
After avoiding cleaning up the input handling for months, I decided the to just make a major refactor of the code base to make it easier to work with. 

My major realization was that I had waaaay too many Systems. Even though each one had relatively little code, it was very brittle and made it hard for the Systems to interact. So I've decided to reduce the number of main systems down to 5. There are additional support Systems that don't really change (and don't really interact with the main systems.) 

This build is broken and doesn't compile. I was in the middle of working on LogicSystem when I had to stop working, but wanted to check it in anyways.